requirements.txt
schedule
pystray
Misskey.py
